[PARA-748] Drop grunt-contrib-yuidoc
This is causing errors when running npm i.  While a downgrade of node/npm could
solve the issue as well, no one is actively using these yui docs right now, and
we have everything in place should we want to regerenate them in the future.

npm Github issue relating to the error:

https://github.com/npm/cli/issues/4992#issuecomment-1152925590